Closes #2 and #4
The time isn't the same no matter what, but a faster cascade time addresses #4 pretty well. It still looks nice and you're not waiting around to get information. There aren't any stops that will have more than ten or so departures, so this speed should be fine. As for #2, it's a pretty simple if statement and some css that applies to the h2 text it inserts.