Some README updates
I also reordered the configuration options to be the same in the two
places in the JS and the README.

Finally, I found an undocumented `logo` option that @sherson says nobody
is using. For some reason I can't put my finger on, my security senses
are tingling, so I removed it.