Use local time fields
My understanding of the new local time fields is that they will remain consistent during a service day, even if that service day includes a time change.

Therefore, we can switch to using the local time fields, and discard our logic which we had constructed to deal with the feed error. Closes #9.

Correct me if I'm wrong.